Add script to compute number of predicted hirings per company per rome
For the PSE school project, they need to have the number of predicted hirings, per siret, per rome.
In the database, we only have the number of total hirings for all romes for one company.
We have to do a little trick with this script, which will find it for each rome.
It is inspired by the create index script.